+++ src/main/java/org/cyclops/integratedtunnels/part/aspect/IEnergyTarget.java
@@ -21,11 +21,11 @@ public interface IEnergyTarget extends IChanneledTarget<IEnergyNetwork, Long> {
 
     public IIngredientComponentStorage<Long, Boolean> getStorage();
 
-    public int getAmount();
+    public long getAmount();
 
     public boolean isExactAmount();
 
-    public static IEnergyTarget ofTile(PartTarget partTarget, IAspectProperties properties, int amount) {
+    public static IEnergyTarget ofTile(PartTarget partTarget, IAspectProperties properties, long amount) {
         PartPos center = partTarget.getCenter();
         PartPos target = partTarget.getTarget();
         INetwork network = IChanneledTarget.getNetworkChecked(center);
@@ -35,7 +35,7 @@ public interface IEnergyTarget extends IChanneledTarget<IEnergyNetwork, Long> {
     }
 
     public static IEnergyTarget ofEntity(PartTarget partTarget, @Nullable Entity entity,
-                                                          IAspectProperties properties, int amount) {
+                                                          IAspectProperties properties, long amount) {
         PartPos center = partTarget.getCenter();
         PartPos target = partTarget.getTarget();
         INetwork network = IChanneledTarget.getNetworkChecked(center);
